bitkeeper revision 1.1159.212.133 (4208f70bysX-Xugo-Q2_ptciUVbe_g)
Use size-independent '%p' format string in preference to '%08lx'.

+++ xen/include/asm-x86/x86_64/regs.h
@@ -9,18 +9,18 @@ struct xen_regs
     u64 r14;
     u64 r13;
     u64 r12;
-    union { u64 rbp; u32 ebp; } __attribute__ ((packed));
-    union { u64 rbx; u32 ebx; } __attribute__ ((packed));
+    union { u64 rbp; u64 ebp; } __attribute__ ((packed));
+    union { u64 rbx; u64 ebx; } __attribute__ ((packed));
     /* NB. Above here is C callee-saves. */
     u64 r11;
     u64 r10;	
     u64 r9;
     u64 r8;
-    union { u64 rax; u32 eax; } __attribute__ ((packed));
-    union { u64 rcx; u32 ecx; } __attribute__ ((packed));
-    union { u64 rdx; u32 edx; } __attribute__ ((packed));
-    union { u64 rsi; u32 esi; } __attribute__ ((packed));
-    union { u64 rdi; u32 edi; } __attribute__ ((packed));
+    union { u64 rax; u64 eax; } __attribute__ ((packed));
+    union { u64 rcx; u64 ecx; } __attribute__ ((packed));
+    union { u64 rdx; u64 edx; } __attribute__ ((packed));
+    union { u64 rsi; u64 esi; } __attribute__ ((packed));
+    union { u64 rdi; u64 edi; } __attribute__ ((packed));
     u32 error_code;
     u32 entry_vector;
     union { u64 rip; u64 eip; } __attribute__ ((packed));
